feat: lock

+++ subscribe.go
@@ -3,7 +3,7 @@ package deepcoin
 type SendTopicAction struct {
 	Action      string `json:"action"`
 	FilterValue string `json:"filterValue"`
-	LocalNo     int    `json:"localNo"`
-	ResumeNo    int    `json:"resumeNo"`
+	LocalNo     int64  `json:"localNo"`
+	ResumeNo    int64  `json:"resumeNo"`
 	TopicID     string `json:"topicID"`
 }
